Refactoring(modals): separando partes dos modals em subpastas utils
