+++ src/bot/modals/compet/registerTalks/registerTalksModal.ts
@@ -8,16 +8,12 @@ import { getCompetTalksRegistration } from "@/bot/utils/googleAPI/getCompetTalks
 import { env } from "@/env";
 import { formatarData } from "@/bot/utils/formatting/formatarData";
 import { uploadToFolder } from "@/bot/utils/googleAPI/googleDrive";
-import { CertificatesType } from "@/api/modules/certificados/entities/certificados.entity"
-import { ExtendedModalInteraction } from "@/bot/typings/Modals"
 import { PythonShellError } from "python-shell";
 import { PythonVenvNotActivatedError } from "@/bot/errors/pythonVenvNotActivatedError";
 import { makeErrorEmbed } from "@/bot/utils/embed/makeErrorEmbed";
 import { makeSuccessEmbed } from "@/bot/utils/embed/makeSuccessEmbed";
-import { validateDriveLink } from "@/bot/utils/googleAPI/updateCompetTalks";
-import { Either, left, right } from "@/api/@types/either";
-import { InvalidDriveLinkError } from "@/bot/errors/invalidDriveLinkError";
-import { FetchReponseError } from "@/bot/errors/fetchReponseError";
+import { extractInputData } from "./utils/extractInputData";
+import { createCertificatesInDatabase } from "./utils/createCertificatesInDB";
 
 // import { submitToAutentique } from "@/bot/utils/autentiqueAPI";
 
@@ -65,8 +61,10 @@ export default new Modal({
 
         if (link) {
             const createCertificatesInDatabaseResponse = await createCertificatesInDatabase({
-                compbio: true, compet_talks: true,
-                data, link, listaNomes, titulo
+                body: {
+                    compbio: true, compet_talks: true,
+                    data, link, listaNomes, titulo
+                }
             })
             if (createCertificatesInDatabaseResponse.isLeft())
                 return await interaction.editReply({
@@ -173,92 +171,5 @@ export default new Modal({
         })
     }
 });
-
-interface ExtractInputDataRequest {
-    interaction: ExtendedModalInteraction,
-    inputFields: TextInputComponentData[]
-}
-
-function extractInputData({ inputFields, interaction }: ExtractInputDataRequest): ExtractInputDataResponse {
-    const customIds = inputFields.map((field) => field.customId || "");
-    const input_data = customIds.map(i => ({ [i]: interaction.fields.getTextInputValue(i) }));
-
-    interface InputFieldsRequest {
-        titulo: string,
-        data_new: string
-        email_assinante: string,
-        nome_assinante: string,
-        minutos_totais: number
-        link: string
-    }
-
-    const { /*email_assinante,*/ link, minutos_totais, /*nome_assinante,*/ titulo, data_new }: InputFieldsRequest = Object.assign({}, ...input_data, { data_inicio: new Date().toISOString() });
-
-    const minutos_input = minutos_totais as number;
-    const timing: { horas: unknown; minutos: unknown } = {
-        horas: Math.trunc(minutos_input / 60),
-        minutos: minutos_input % 60,
-    };
-    const { horas, minutos } = {
-        horas: timing.horas as string,
-        minutos: timing.minutos as string,
-    };
-
-    const parts = data_new.split("-"); // Split the string into parts
-    const day = parseInt(parts[0], 10); // Parse day as an integer
-    const month = parseInt(parts[1], 10) - 1; // Parse month as an integer (months are 0-indexed)
-    const year = parseInt(parts[2], 10); // Parse year as an integer
-
-    const data = new Date(year, month, day);
-
-    return { horas, link, minutos, titulo, data }
-}
-
-interface ExtractInputDataResponse {
-    horas: string,
-    minutos: string,
-    link?: string,
-    email_assinante?: string
-    nome_assinante?: string
-    titulo: string
-    data: Date
-}
-
-type CreateCertificatesInDatabase = Either<
-    { error: InvalidDriveLinkError },
-    { sucess: boolean }
->
-
-/**
- * @author Henrique de Paula Rodrigues
- * @description Cadastra certificados no banco de dados
- */
-async function createCertificatesInDatabase(body: CertificatesType): Promise<CreateCertificatesInDatabase> {
-    if (!validateDriveLink(body.link))
-        return left({
-            error: new InvalidDriveLinkError()
-        })
-
-    const url = `${env.HOST}certificados/`
-    const options = {
-        method: "post",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(body),
-    }
-
-    // TODO: enviar email para a lista de participantes informando que o certificado está disponivel
-    const response = await fetch(url, options);
-    if (!(200 <= response.status && response.status < 300)) {
-
-        const { code, message, status }: { code: number; message: string; status: number } = await response.clone().json();
-        return left({
-            error: new FetchReponseError({ code, status, message })
-        })
-    }
-
-    return right({
-        sucess: true
-    })
-}
 
 export { registerTalksModal };
